+++ lib/tessa.rb
@@ -10,7 +10,6 @@ require "tessa/asset"
 require "tessa/asset_change"
 require "tessa/asset_change_set"
 require "tessa/controller_helpers"
-require "tessa/failure"
 require "tessa/model"
 require "tessa/rack_upload_proxy"
 require "tessa/upload"
